+++ config/secrets.go
@@ -9,7 +9,6 @@ const secretsSettingsName = "secrets"
 
 type Secrets struct {
 	SessionKey []byte
-	CsrfKey    []byte
 }
 
 func LoadSecrets(store Store) (*Secrets, error) {
@@ -26,15 +25,6 @@ func LoadSecrets(store Store) (*Secrets, error) {
 			return nil, err
 		}
 		s.SessionKey = newKey
-		dirty = true
-	}
-
-	if s.CsrfKey == nil || len(s.CsrfKey) < 32 {
-		newKey := make([]byte, 32)
-		if _, err := io.ReadFull(rand.Reader, newKey); err != nil {
-			return nil, err
-		}
-		s.CsrfKey = newKey
 		dirty = true
 	}