+++ handlers_search.go
@@ -1,6 +1,7 @@
 package main
 
 import (
+	"log"
 	"net/http"
 )
 
@@ -10,6 +11,12 @@ type SearchRequest interface {
 
 func SearchHandler(templates *Templates, backend SearchRequest) http.HandlerFunc {
 	return func(w http.ResponseWriter, r *http.Request) {
+		if err := r.ParseForm(); err != nil {
+			log.Printf("Error parsing form: %v", err)
+			w.WriteHeader(http.StatusInternalServerError)
+			return
+		}
+
 		pattern := r.FormValue("pattern")
 		var results []SearchResult
 		if pattern != "" {